requirements.txt
cython
numpy
httplib2
django
django-cors-headers
gunicorn
gevent
geojson
iso8601
django-jquery
pinax-theme-bootstrap
django_forms_bootstrap
psycopg2
requests
pyyaml
django-taggit
django-taggit-templatetags
django-leaflet
pymemcache
bleach
markdown
unidecode
git+https://github.com/geodashio/geodash-framework-django.git#egg=geodash-framework-django
